# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Arm Bot Back App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /hello:
    # binds swagger app logic to a route
    x-swagger-router-controller: hello_world
    get:
      description: Returns 'Hello' to the caller
      # used as the method name of the controller
      operationId: hello
      parameters:
        - name: name
          in: query
          description: The name of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /park/lotation:
    # binds a127 app logic to a route
    x-swagger-router-controller: park
    get:
      description: Returns the number of vehicles at the park
      # used as the method name of the controller
      operationId: lotation
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ParkLotationSuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /park/checkin:
    # binds a127 app logic to a route
    x-swagger-router-controller: park
    get:
      description: Adds 1 to the park lotation
      # used as the method name of the controller
      operationId: checkin
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ParkCheckinSucessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /park/checkout:
    # binds a127 app logic to a route
    x-swagger-router-controller: park
    get:
      description: Decreases 1 to the park lotation
      # used as the method name of the controller
      operationId: checkout
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ParkCheckoutSucessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
  ParkLotationSuccessResponse:
    required:
      - vehicles
    properties:
      vehicles:
        type: integer
  ParkCheckinSucessResponse:
    required:
      - vehicles
    properties:
      vehicles:
        type: integer
  ParkCheckoutSucessResponse:
    required:
      - vehicles
    properties:
      vehicles:
        type: integer
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string